+++ src/main/java/domain/BasketItem.java
@@ -1,16 +1,18 @@
 package domain;
 
+import java.math.BigDecimal;
+
 public class BasketItem {
 
     private Item item;
-    private Double price;
+    private BigDecimal price;
 
-    public BasketItem(Item item, Double price) {
+    public BasketItem(Item item, BigDecimal price) {
         this.item = item;
         this.price = price;
     }
 
-    public Double price() {
+    public BigDecimal price() {
         return price;
     }
 }
